fix the Linux build: WORD is not portable
GLB_ValidItem borrowed its casts from the ASSERT lines just below it, but
ASSERT is ((void)0) so those casts never actually compile - the Windows-only
WORD typedef had simply never been instantiated there. filenum/itemnum are
uint16_t and items is int, so compare as int.

+++ src/glbapi.cpp
@@ -877,10 +877,10 @@ GLB_ValidItem(
 
 	itm.handle = MOD_Resolve(handle);
 
-	if (itm.id.filenum >= (WORD)num_glbs)
+	if ((int)itm.id.filenum >= num_glbs)
 		return 0;
 
-	if (itm.id.itemnum >= (WORD)filedesc[itm.id.filenum].items)
+	if ((int)itm.id.itemnum >= filedesc[itm.id.filenum].items)
 		return 0;
 
 	return 1;
